Refactor ManageListingPage route config.
