+++ src/containers/pageDataLoadingAPI.js
@@ -5,6 +5,7 @@ import { setInitialValues as CheckoutPageInitialValues } from './CheckoutPage/Ch
 import { loadData as EditListingPageLoader } from './EditListingPage/EditListingPage.duck';
 import { loadData as InboxPageLoader } from './InboxPage/InboxPage.duck';
 import { loadData as ListingPageLoader } from './ListingPage/ListingPage.duck';
+import { loadData as ManageListingsPageLoader } from './ManageListingsPage/ManageListingsPage.duck';
 import { loadData as ProfilePageLoader } from './ProfilePage/ProfilePage.duck';
 import { loadData as SearchPageLoader } from './SearchPage/SearchPage.duck';
 import {
@@ -25,6 +26,9 @@ const getPageDataLoadingAPI = () => {
     },
     ListingPage: {
       loadData: ListingPageLoader,
+    },
+    ManageListingsPage: {
+      loadData: ManageListingsPageLoader,
     },
     ProfilePage: {
       loadData: ProfilePageLoader,
